requirements.txt
requests
openai-whisper
python-telegram-bot
ffmpeg-python
python-socketio[asyncio_client]
python-socketio[asyncio_server]
aiohttp
aiohttp-cors 